fix: moved enums to a common enum folder. Removed duplicate Faction enum. Added CrossOrigin annotation to endpoints that were missing them.
