+++ src/main/java/experis/humansvszombies/hvz/models/enums/UserType.java
@@ -1,0 +1,6 @@
+package experis.humansvszombies.hvz.models.enums;
+
+public enum UserType {
+    ADMINISTRATOR,
+    PLAYER
+}
