+++ src/main/java/experis/humansvszombies/hvz/models/tables/Mission.java
@@ -9,8 +9,9 @@ import javax.persistence.ManyToOne;
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
-import experis.humansvszombies.hvz.models.tables.enums.Faction;
-import experis.humansvszombies.hvz.models.tables.enums.MissionState;
+
+import experis.humansvszombies.hvz.models.enums.Faction;
+import experis.humansvszombies.hvz.models.enums.MissionState;
 
 import java.sql.Timestamp;
