--- src/main/java/experis/humansvszombies/hvz/models/tables/enums/Faction.java
@@ -1,7 +1,0 @@
-package experis.humansvszombies.hvz.models.tables.enums;
-
-public enum Faction {
-    HUMAN,
-    ZOMBIE
-}
-
