--- src/main/java/experis/humansvszombies/hvz/models/tables/enums/UserType.java
@@ -1,6 +1,0 @@
-package experis.humansvszombies.hvz.models.tables.enums;
-
-public enum UserType {
-    ADMINSTRATOR,
-    PLAYER
-}
